Migrated intervals::absolute::{finite_bound, start_bound, end_bound} and their examples to jiff
